## Context

It reduces memory allocation by 80% and increases throughput by 20% as well.

Here are the benchmark results on my Apple Air M1 :)

```
# JMH version: 1.36
# VM version: JDK 17.0.1, OpenJDK 64-Bit Server VM, 17.0.1+12
# VM invoker: ...java/17.0.1-tem/bin/java


# Before
Benchmark                                                      Mode  Cnt            Score    Error   Units
JaroWinklerSimilarityBenchmark.benchmark                      thrpt   15            0.082 ±  0.002   ops/s
JaroWinklerSimilarityBenchmark.benchmark:·gc.alloc.rate       thrpt   15         1006.070 ± 27.086  MB/sec
JaroWinklerSimilarityBenchmark.benchmark:·gc.alloc.rate.norm  thrpt   15  12926893363.733 ± 82.405    B/op
JaroWinklerSimilarityBenchmark.benchmark:·gc.count            thrpt   15          277.000           counts
JaroWinklerSimilarityBenchmark.benchmark:·gc.time             thrpt   15          168.000               ms

# After
Benchmark                                                      Mode  Cnt           Score     Error   Units
JaroWinklerSimilarityBenchmark.benchmark                      thrpt   15           0.099 ±   0.002   ops/s
JaroWinklerSimilarityBenchmark.benchmark:·gc.alloc.rate       thrpt   15         194.065 ±   3.370  MB/sec
JaroWinklerSimilarityBenchmark.benchmark:·gc.alloc.rate.norm  thrpt   15  2050897564.267 ± 127.669    B/op
JaroWinklerSimilarityBenchmark.benchmark:·gc.count            thrpt   15          55.000            counts
JaroWinklerSimilarityBenchmark.benchmark:·gc.time             thrpt   15          36.000                ms

```

## Checklist
- [x] Change meets or does not compromise the [Baseline Security Requirements](https://transferwise.atlassian.net/wiki/spaces/EKB/pages/434929973/Baseline+Security+Requirements)